Split driver candidates to their own file
