+++ collector/lib/GetKernelObject.cpp
@@ -138,8 +138,8 @@ bool DownloadKernelObject(const std::string& hostname, const Json::Value& tls_co
   return false;
 }
 
-bool GetKernelObject(const std::string& hostname, const Json::Value& tls_config, const std::string& kernel_module, const std::string& module_path, bool verbose) {
-  std::string expected_path = kKernelModulesDir + "/" + kernel_module;
+bool GetKernelObject(const std::string& hostname, const Json::Value& tls_config, const DriverCandidate& candidate, const std::string& module_path, bool verbose) {
+  std::string expected_path = candidate.getPath() + "/" + candidate.getName();
   std::string expected_path_compressed = expected_path + ".gz";
   struct stat sb;
 
@@ -176,11 +176,11 @@ bool GetKernelObject(const std::string& hostname, const Json::Value& tls_config,
               std::ostreambuf_iterator<char>(output_file));
   }
   // Otherwise there is no module in local storage, so we should download it.
-  else {
-    CLOG(INFO) << "Local storage does not contain " << kernel_module;
+  else if (candidate.isDownloadable()) {
+    CLOG(INFO) << "Attempting to download " << candidate.getName();
     std::string downloadPath = module_path + ".gz";
-    if (!DownloadKernelObject(hostname, tls_config, kernel_module, downloadPath, verbose)) {
-      CLOG(WARNING) << "Unable to download kernel object " << kernel_module << " to " << downloadPath;
+    if (!DownloadKernelObject(hostname, tls_config, candidate.getName(), downloadPath, verbose)) {
+      CLOG(WARNING) << "Unable to download kernel object " << candidate.getName() << " to " << downloadPath;
       return false;
     }
 
@@ -192,6 +192,9 @@ bool GetKernelObject(const std::string& hostname, const Json::Value& tls_config,
       return false;
     }
     CLOG(INFO) << "Successfully downloaded and decompressed " << module_path;
+  } else {
+    CLOG(WARNING) << "Local storage does not contain " << candidate.getName();
+    return false;
   }
 
   if (chmod(module_path.c_str(), 0444)) {
