+++ collector/test/DriverCandidatesTest.cpp
@@ -24,8 +24,8 @@ You should have received a copy of the GNU General Public License along with thi
 #include <gmock/gmock-actions.h>
 #include <gmock/gmock-spec-builders.h>
 
+#include "DriverCandidates.cpp"
 #include "HostInfo.h"
-#include "Utility.cpp"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
